refactor: 테스트용 상수 prefix 추가
