+++ src/test/java/com/gdschongik/gdsc/domain/email/application/UnivEmailVerificationServiceTest.java
@@ -1,6 +1,6 @@
 package com.gdschongik.gdsc.domain.email.application;
 
-import static com.gdschongik.gdsc.global.common.constant.MemberConstant.*;
+import static com.gdschongik.gdsc.global.common.constant.TestMemberConstant.*;
 import static com.gdschongik.gdsc.global.exception.ErrorCode.*;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
