+++ src/test/java/com/gdschongik/gdsc/domain/event/application/EventParticipationServiceTest.java
@@ -3,8 +3,8 @@ package com.gdschongik.gdsc.domain.event.application;
 import static com.gdschongik.gdsc.domain.event.domain.AfterPartyAttendanceStatus.*;
 import static com.gdschongik.gdsc.domain.event.domain.UsageStatus.*;
 import static com.gdschongik.gdsc.domain.member.domain.Department.*;
-import static com.gdschongik.gdsc.global.common.constant.EventConstant.*;
-import static com.gdschongik.gdsc.global.common.constant.MemberConstant.*;
+import static com.gdschongik.gdsc.global.common.constant.TestEventConstant.*;
+import static com.gdschongik.gdsc.global.common.constant.TestMemberConstant.*;
 import static com.gdschongik.gdsc.global.exception.ErrorCode.*;
 import static org.assertj.core.api.Assertions.*;
