+++ src/test/java/com/gdschongik/gdsc/domain/event/application/EventServiceTest.java
@@ -1,6 +1,6 @@
 package com.gdschongik.gdsc.domain.event.application;
 
-import static com.gdschongik.gdsc.global.common.constant.EventConstant.*;
+import static com.gdschongik.gdsc.global.common.constant.TestEventConstant.*;
 import static com.gdschongik.gdsc.global.exception.ErrorCode.*;
 import static org.assertj.core.api.Assertions.*;
