+++ src/test/java/com/gdschongik/gdsc/domain/event/domain/ParticipantRoleTest.java
@@ -1,7 +1,7 @@
 package com.gdschongik.gdsc.domain.event.domain;
 
 import static com.gdschongik.gdsc.domain.member.domain.Department.*;
-import static com.gdschongik.gdsc.global.common.constant.MemberConstant.*;
+import static com.gdschongik.gdsc.global.common.constant.TestMemberConstant.*;
 import static com.gdschongik.gdsc.global.exception.ErrorCode.*;
 import static org.assertj.core.api.Assertions.*;
