+++ src/test/java/com/gdschongik/gdsc/domain/study/domain/StudyFactoryTest.java
@@ -1,6 +1,6 @@
 package com.gdschongik.gdsc.domain.study.domain;
 
-import static com.gdschongik.gdsc.global.common.constant.StudyConstant.*;
+import static com.gdschongik.gdsc.global.common.constant.TestStudyConstant.*;
 import static org.assertj.core.api.Assertions.*;
 
 import com.gdschongik.gdsc.domain.member.domain.Member;
