+++ src/test/java/com/gdschongik/gdsc/global/common/constant/TestCouponConstant.java
@@ -1,7 +1,7 @@
 package com.gdschongik.gdsc.global.common.constant;
 
-public class CouponConstant {
+public class TestCouponConstant {
     public static final String COUPON_NAME = "테스트 쿠폰 이름";
 
-    private CouponConstant() {}
+    private TestCouponConstant() {}
 }
