+++ src/test/java/com/gdschongik/gdsc/global/common/constant/TestOrderConstant.java
@@ -1,10 +1,10 @@
 package com.gdschongik.gdsc.global.common.constant;
 
-public class OrderConstant {
+public class TestOrderConstant {
 
     public static final String ORDER_NANO_ID = "HnbMWoSZRq3qK1W3tPXCW";
     public static final String ORDER_PAYMENT_KEY = "testPaymentKey";
     public static final String ORDER_CANCEL_REASON = "테스트 주문 취소 사유";
 
-    private OrderConstant() {}
+    private TestOrderConstant() {}
 }
