+++ src/test/java/com/gdschongik/gdsc/global/common/constant/TestRecruitmentConstant.java
@@ -6,7 +6,7 @@ import com.gdschongik.gdsc.domain.recruitment.domain.RoundType;
 import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
-public class RecruitmentConstant {
+public class TestRecruitmentConstant {
     // 1차 모집 상수
     public static final String RECRUITMENT_ROUND_NAME = "2024학년도 1학기 1차 모집";
     public static final LocalDateTime START_DATE = LocalDateTime.of(2024, 3, 2, 0, 0);
@@ -26,5 +26,5 @@ public class RecruitmentConstant {
     public static final LocalDateTime ROUND_TWO_END_DATE = LocalDateTime.of(2024, 3, 10, 0, 0);
     public static final Period ROUND_TWO_START_TO_END_PERIOD = Period.of(ROUND_TWO_START_DATE, ROUND_TWO_END_DATE);
 
-    private RecruitmentConstant() {}
+    private TestRecruitmentConstant() {}
 }
