+++ src/test/java/com/gdschongik/gdsc/global/common/constant/TestTemporalConstant.java
@@ -4,7 +4,7 @@ import com.gdschongik.gdsc.domain.common.model.SemesterType;
 import com.gdschongik.gdsc.domain.common.vo.Semester;
 import java.time.LocalDateTime;
 
-public class TemporalConstant {
+public class TestTemporalConstant {
     // academic year
     public static final Integer ACADEMIC_YEAR = 2024;
 
@@ -14,5 +14,5 @@ public class TemporalConstant {
     public static final LocalDateTime SEMESTER_END_DATE = LocalDateTime.of(2024, 8, 31, 0, 0);
     public static final Semester SEMESTER = Semester.of(ACADEMIC_YEAR, SEMESTER_TYPE);
 
-    private TemporalConstant() {}
+    private TestTemporalConstant() {}
 }
